requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
html2markdown==0.1.7
isort==4.3.17
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
pylint==2.3.1
PyYAML==5.1
six==1.12.0
soupsieve==1.9.1
typed-ast==1.3.4
wrapt==1.11.1